+++ src/main/kotlin/org/careerseekers/cseventsservice/utils/StatisticsStorage.kt
@@ -1,7 +1,6 @@
 package org.careerseekers.cseventsservice.utils
 
 import java.time.LocalDateTime
-import java.util.concurrent.atomic.AtomicInteger
 import java.util.concurrent.atomic.AtomicLong
 
 object StatisticsStorage {
@@ -23,10 +22,10 @@ object StatisticsStorage {
     var lastDocumentUpload: LocalDateTime? = null
         private set
 
-    var eventsCount: AtomicInteger = AtomicInteger(0)
+    var eventsCount: AtomicLong = AtomicLong(0)
         private set
 
-    var verifiedEventsCount: AtomicInteger = AtomicInteger(0)
+    var verifiedEventsCount: AtomicLong = AtomicLong(0)
         private set
 
     fun setPlatformsCount(platformsCount: Long) = apply { this.platformsCount.set(platformsCount) }
@@ -37,6 +36,6 @@ object StatisticsStorage {
 
     fun setDirectionDocsCount(count: Long) = apply { this.directionDocsCount.set(count) }
     fun setLastDocumentUpload(lastDocumentUpload: LocalDateTime?) = apply { this.lastDocumentUpload = lastDocumentUpload }
-    fun setEventsCount(count: Int) = apply { this.eventsCount.set(count) }
-    fun setVerifiedEventsCount(count: Int) = apply { this.verifiedEventsCount.set(count) }
+    fun setEventsCount(count: Long) = apply { this.eventsCount.set(count) }
+    fun setVerifiedEventsCount(count: Long) = apply { this.verifiedEventsCount.set(count) }
 }
